+++ packages/screenshoter/CHANGELOG.md
@@ -4,6 +4,15 @@
 ## Unreleased
 
 
+## 3.2.3 - 2021/8/13
+
+- remove base64 sanitizing
+
+## 3.2.2 - 2021/8/13
+
+- add `withStatusBar` capability property to take app and full app screenshots with status bar
+- fix ios web screenshots
+
 ## 3.2.1 - 2021/8/8
 
 - no changes